Added round,roundN,roundDigitsN  and abs functions
There is already ceil and floor, but natural rounding was missing. Added.
Added also with N version, which allows to set number of *fractional*
digits we are interested in.
For compeltness, also roundDigitsN is added which rounds simply to number
of digits, and length functions, telling how much digits number have